requirements.txt
torch
numpy
google-generativeai
pyyaml
langchain-ollama
langchain-core
pymongo
colorama
requests>=2.28.0
Pillow>=9.0.0
numpy
kornia 
opencv-python